requirements.txt
pytest == 8.4.1
requests == 2.32.4
allure-pytest == 2.14.3
Faker == 37.4.2
